PR: MPH-21
When help:effective-pom is called from within a pom,
only print out that pom's effective pom.
Added check to only dump all poms if the ${project} is the head of
the reactor list to enforce proper aggregator behaviour.

TODO: maybe this needs an update: find the current project
in the reactor projects, then dump that pom, and find all
modules recursively and dump those, so you can dump a sub-tree
of the reactor, when you run the mojo from the lifecycle.



git-svn-id: https://svn.apache.org/repos/asf/maven/plugins/trunk@497785 13f79535-47bb-0310-9956-ffa450edef68