[MPH-99] Evaluate has no output in quiet mode
Adding the 'output' parameter to the evaluate Mojo. When used, along with an expression, the goal will write the result to the specified file instead of logging to the console. If no expression is specified, 'output' will be ignored (since the plugin will prompt for the expression, it is intended that the result is written to the console). The 'quiet' mode is documented to only output error messages, so having the result in quiet mode would break the usage of this parameter. This closes #17.

git-svn-id: https://svn.apache.org/repos/asf/maven/plugins/trunk@1765191 13f79535-47bb-0310-9956-ffa450edef68