[MPH-105] Effective pom aggregation is not triggered
When the plugin is invoked from the command-line, always show all effective POMs for all projects in the reactor. When it is invoked from the lifecycle, only show all effective POMs for the head project in the reactor (see commit 497785 and resolution of MPH-21); otherwise, show the effective POM for the current project only.

git-svn-id: https://svn.apache.org/repos/asf/maven/plugins/trunk@1769446 13f79535-47bb-0310-9956-ffa450edef68