[MPH-168] - Only write individual effective files if output is specified
When the user only specifies the individual flag, it will create individual effective POMs and print it to standard out. When the output parameter is specified, it is used to write the ouput to files, relative to the (sub)project the effective POM is for. The output parameter is parsed as a File so contains the full path. We replace the basedir of that path to get the actual location where the effective POM should be written to. That would fail in case of a relative path going up the tree. Therefore, if the basedir cannot be found, we take the name of the output path and use only that.